pyproject.toml
[tool.poetry]
name = "codestats"
version = "0.1.0"
description = ""
authors = ["Jakub Zárybnický <jakub@zarybnicky.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.28.2"
pandas = "^2.1.3"
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
gitpython = "^3.1.40"
# git-pandas = {git = "https://github.com/wdm0006/git-pandas"}
sqlacodegen = "3.0.0rc3"
python-dotenv = "^1.0.0"
humanize = "^4.9.0"
rich = "^13.7.0"
duckdb = "^0.9.2"

[tool.pyright]
include = ['.']

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
